small fix, need to find a way to use all types, though

+++ python/src/random.cpp
@@ -7,10 +7,10 @@ namespace py = pybind11;
 void init_random(py::module& m)
 {
     m.def("random_uniform",
-          py::overload_cast<int, int>(&duda::random_uniform<double>),
+          py::overload_cast<int, int>(&duda::random_uniform<float>),
           "random uniform for a matrix");
 
     m.def("random_uniform",
-          py::overload_cast<int>(&duda::random_uniform<double>),
+          py::overload_cast<int>(&duda::random_uniform<float>),
           "random uniform for a vector");
 }
